app/frozen_requirements.txt
anyio==3.4.0
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
click==8.0.3
cryptography==36.0.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
h11==0.12.0
httptools==0.2.0
idna==3.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pymongo==3.12.1
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.26.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
